fix : 수정

+++ server/src/main/java/gohome/dailydaily/domain/product/dto/SelectDto.java
@@ -10,6 +10,6 @@ public class SelectDto {
     private Long productId;
     private String title;
     private String main;
-    private Float score;
+    private Integer score;
 
 }
